Feat. 한줄씩 완성

+++ Game Framework/SDL Game Framework/Framework/Csv.c
@@ -18,7 +18,7 @@ void readFileToBuffer(const char* filename)
 	}
 
 	fseek(fp, 0, SEEK_END);
-	long fileSize = ftell(fp);
+	long long fileSize = ftell(fp);
 
 	s_Buffer = malloc(fileSize + 1);
 	memset(s_Buffer, 0, fileSize + 1);
@@ -58,7 +58,7 @@ void CreateCsvFile(CsvFile* csvFile, const char* filename)
 	s_BufferPointer = s_Buffer;
 	while (*s_BufferPointer != '\0')
 	{
-		int row = csvFile->RowCount;
+		int32 row = csvFile->RowCount;
 
 		int commaCount = 0;
 		const char* lineStart = s_BufferPointer;
